Renamed files to match with python examples and added a new example
