+++ examples/src/04_h264_encoding.cpp
