+++ examples/src/08_rgb_mobilenet.cpp
