+++ examples/src/16_device_queue_event.cpp
