+++ examples/src/20_image_manip_warp.cpp
