+++ examples/src/24_opencv_support.cpp
