+++ examples/src/25_system_information.cpp
