+++ examples/src/26_2_spatial_mobilenet_mono.cpp
@@ -12,62 +12,64 @@ static const std::vector<std::string> labelMap = {"background", "aeroplane", "bi
                                                   "motorbike",  "person",    "pottedplant", "sheep", "sofa",        "train",  "tvmonitor"};
 
 static bool syncNN = true;
+static bool flipRectified = true;
 
-dai::Pipeline createNNPipeline(std::string nnPath) {
+dai::Pipeline createPipeline(std::string nnPath) {
     dai::Pipeline p;
 
-    // create nodes
-    auto colorCam = p.create<dai::node::ColorCamera>();
+    auto xlinkOut = p.create<dai::node::XLinkOut>();
     auto spatialDetectionNetwork = p.create<dai::node::MobileNetSpatialDetectionNetwork>();
+    auto imageManip = p.create<dai::node::ImageManip>();
+
+    auto nnOut = p.create<dai::node::XLinkOut>();
+    auto depthRoiMap = p.create<dai::node::XLinkOut>();
+    auto xoutDepth = p.create<dai::node::XLinkOut>();
+
+    xlinkOut->setStreamName("preview");
+    nnOut->setStreamName("detections");
+    depthRoiMap->setStreamName("boundingBoxDepthMapping");
+    xoutDepth->setStreamName("depth");
+
+    imageManip->initialConfig.setResize(300, 300);
+    // The NN model expects BGR input. By default ImageManip output type would be same as input (gray in this case)
+    imageManip->initialConfig.setFrameType(dai::RawImgFrame::Type::BGR888p);
+
     auto monoLeft = p.create<dai::node::MonoCamera>();
     auto monoRight = p.create<dai::node::MonoCamera>();
     auto stereo = p.create<dai::node::StereoDepth>();
-
-    // create xlink connections
-    auto xoutRgb = p.create<dai::node::XLinkOut>();
-    auto xoutNN = p.create<dai::node::XLinkOut>();
-    auto xoutBoundingBoxDepthMapping = p.create<dai::node::XLinkOut>();
-    auto xoutDepth = p.create<dai::node::XLinkOut>();
-
-    xoutRgb->setStreamName("preview");
-    xoutNN->setStreamName("detections");
-    xoutBoundingBoxDepthMapping->setStreamName("boundingBoxDepthMapping");
-    xoutDepth->setStreamName("depth");
-
-    colorCam->setPreviewSize(300, 300);
-    colorCam->setResolution(dai::ColorCameraProperties::SensorResolution::THE_1080_P);
-    colorCam->setInterleaved(false);
-    colorCam->setColorOrder(dai::ColorCameraProperties::ColorOrder::BGR);
-
     monoLeft->setResolution(dai::MonoCameraProperties::SensorResolution::THE_400_P);
     monoLeft->setBoardSocket(dai::CameraBoardSocket::LEFT);
     monoRight->setResolution(dai::MonoCameraProperties::SensorResolution::THE_400_P);
     monoRight->setBoardSocket(dai::CameraBoardSocket::RIGHT);
 
-    /// setting node configs
+    // StereoDepth
     stereo->setOutputDepth(true);
+    stereo->setOutputRectified(true);
     stereo->setConfidenceThreshold(255);
-
-    spatialDetectionNetwork->setBlobPath(nnPath);
-    spatialDetectionNetwork->setConfidenceThreshold(0.5f);
-    spatialDetectionNetwork->input.setBlocking(false);
-    spatialDetectionNetwork->setBoundingBoxScaleFactor(0.5);
-    spatialDetectionNetwork->setDepthLowerThreshold(100);
-    spatialDetectionNetwork->setDepthUpperThreshold(5000);
 
     // Link plugins CAM -> STEREO -> XLINK
     monoLeft->out.link(stereo->left);
     monoRight->out.link(stereo->right);
 
-    // Link plugins CAM -> NN -> XLINK
-    colorCam->preview.link(spatialDetectionNetwork->input);
-    if(syncNN)
-        spatialDetectionNetwork->passthrough.link(xoutRgb->input);
-    else
-        colorCam->preview.link(xoutRgb->input);
+    stereo->rectifiedRight.link(imageManip->inputImage);
 
-    spatialDetectionNetwork->out.link(xoutNN->input);
-    spatialDetectionNetwork->boundingBoxMapping.link(xoutBoundingBoxDepthMapping->input);
+    // testing MobileNet DetectionNetwork
+    spatialDetectionNetwork->setConfidenceThreshold(0.5f);
+    spatialDetectionNetwork->setBoundingBoxScaleFactor(0.7);
+    spatialDetectionNetwork->setDepthLowerThreshold(100);
+    spatialDetectionNetwork->setDepthUpperThreshold(5000);
+
+    spatialDetectionNetwork->setBlobPath(nnPath);
+
+    // Link plugins CAM -> NN -> XLINK
+    imageManip->out.link(spatialDetectionNetwork->input);
+    if(syncNN)
+        spatialDetectionNetwork->passthrough.link(xlinkOut->input);
+    else
+        imageManip->out.link(xlinkOut->input);
+
+    spatialDetectionNetwork->out.link(nnOut->input);
+    spatialDetectionNetwork->boundingBoxMapping.link(depthRoiMap->input);
 
     stereo->depth.link(spatialDetectionNetwork->inputDepth);
     spatialDetectionNetwork->passthroughDepth.link(xoutDepth->input);
@@ -89,7 +91,7 @@ int main(int argc, char** argv) {
     printf("Using blob at path: %s\n", nnPath.c_str());
 
     // Create pipeline
-    dai::Pipeline p = createNNPipeline(nnPath);
+    dai::Pipeline p = createPipeline(nnPath);
 
     // Connect to device with above created pipeline
     dai::Device d(p);
@@ -98,7 +100,7 @@ int main(int argc, char** argv) {
 
     auto preview = d.getOutputQueue("preview", 4, false);
     auto detections = d.getOutputQueue("detections", 4, false);
-    auto xoutBoundingBoxDepthMapping = d.getOutputQueue("boundingBoxDepthMapping", 4, false);
+    auto depthRoiMap = d.getOutputQueue("boundingBoxDepthMapping", 4, false);
     auto depthQueue = d.getOutputQueue("depth", 4, false);
 
     auto startTime = steady_clock::now();
@@ -107,9 +109,18 @@ int main(int argc, char** argv) {
     auto color = cv::Scalar(255, 255, 255);
 
     while(1) {
-        auto imgFrame = preview->get<dai::ImgFrame>();
+        auto inRectifiedRight = preview->get<dai::ImgFrame>();
         auto det = detections->get<dai::SpatialImgDetections>();
         auto depth = depthQueue->get<dai::ImgFrame>();
+
+        counter++;
+        auto currentTime = steady_clock::now();
+        auto elapsed = duration_cast<duration<float>>(currentTime - startTime);
+        if(elapsed > seconds(1)) {
+            fps = counter / elapsed.count();
+            counter = 0;
+            startTime = currentTime;
+        }
 
         auto dets = det->detections;
 
@@ -120,7 +131,7 @@ int main(int argc, char** argv) {
         cv::applyColorMap(depthFrameColor, depthFrameColor, cv::COLORMAP_HOT);
 
         if(!dets.empty()) {
-            auto boundingBoxMapping = xoutBoundingBoxDepthMapping->get<dai::SpatialLocationCalculatorConfig>();
+            auto boundingBoxMapping = depthRoiMap->get<dai::SpatialLocationCalculatorConfig>();
             auto roiDatas = boundingBoxMapping->getConfigData();
 
             for(auto roiData : roiDatas) {
@@ -136,52 +147,51 @@ int main(int argc, char** argv) {
                 cv::rectangle(depthFrameColor, cv::Rect(cv::Point(xmin, ymin), cv::Point(xmax, ymax)), color, cv::FONT_HERSHEY_SIMPLEX);
             }
         }
-        counter++;
-        auto currentTime = steady_clock::now();
-        auto elapsed = duration_cast<duration<float>>(currentTime - startTime);
-        if(elapsed > seconds(1)) {
-            fps = counter / elapsed.count();
-            counter = 0;
-            startTime = currentTime;
-        }
 
-        cv::Mat frame = imgFrame->getCvFrame();
+        cv::Mat rectifiedRight = inRectifiedRight->getCvFrame();
 
-        for(const auto& d : dets) {
-            int x1 = d.xmin * frame.cols;
-            int y1 = d.ymin * frame.rows;
-            int x2 = d.xmax * frame.cols;
-            int y2 = d.ymax * frame.rows;
+        if(flipRectified) cv::flip(rectifiedRight, rectifiedRight, 1);
+
+        for(auto& d : dets) {
+            if(flipRectified) {
+                auto swap = d.xmin;
+                d.xmin = 1 - d.xmax;
+                d.xmax = 1 - swap;
+            }
+            int x1 = d.xmin * rectifiedRight.cols;
+            int y1 = d.ymin * rectifiedRight.rows;
+            int x2 = d.xmax * rectifiedRight.cols;
+            int y2 = d.ymax * rectifiedRight.rows;
 
             int labelIndex = d.label;
             std::string labelStr = to_string(labelIndex);
             if(labelIndex < labelMap.size()) {
                 labelStr = labelMap[labelIndex];
             }
-            cv::putText(frame, labelStr, cv::Point(x1 + 10, y1 + 20), cv::FONT_HERSHEY_TRIPLEX, 0.5, color);
+            cv::putText(rectifiedRight, labelStr, cv::Point(x1 + 10, y1 + 20), cv::FONT_HERSHEY_TRIPLEX, 0.5, color);
             std::stringstream confStr;
             confStr << std::fixed << std::setprecision(2) << d.confidence * 100;
-            cv::putText(frame, confStr.str(), cv::Point(x1 + 10, y1 + 35), cv::FONT_HERSHEY_TRIPLEX, 0.5, color);
+            cv::putText(rectifiedRight, confStr.str(), cv::Point(x1 + 10, y1 + 35), cv::FONT_HERSHEY_TRIPLEX, 0.5, color);
 
             std::stringstream depthX;
             depthX << "X: " << (int)d.spatialCoordinates.x << " mm";
-            cv::putText(frame, depthX.str(), cv::Point(x1 + 10, y1 + 50), cv::FONT_HERSHEY_TRIPLEX, 0.5, color);
+            cv::putText(rectifiedRight, depthX.str(), cv::Point(x1 + 10, y1 + 50), cv::FONT_HERSHEY_TRIPLEX, 0.5, color);
             std::stringstream depthY;
             depthY << "Y: " << (int)d.spatialCoordinates.y << " mm";
-            cv::putText(frame, depthY.str(), cv::Point(x1 + 10, y1 + 65), cv::FONT_HERSHEY_TRIPLEX, 0.5, color);
+            cv::putText(rectifiedRight, depthY.str(), cv::Point(x1 + 10, y1 + 65), cv::FONT_HERSHEY_TRIPLEX, 0.5, color);
             std::stringstream depthZ;
             depthZ << "Z: " << (int)d.spatialCoordinates.z << " mm";
-            cv::putText(frame, depthZ.str(), cv::Point(x1 + 10, y1 + 80), cv::FONT_HERSHEY_TRIPLEX, 0.5, color);
+            cv::putText(rectifiedRight, depthZ.str(), cv::Point(x1 + 10, y1 + 80), cv::FONT_HERSHEY_TRIPLEX, 0.5, color);
 
-            cv::rectangle(frame, cv::Rect(cv::Point(x1, y1), cv::Point(x2, y2)), color, cv::FONT_HERSHEY_SIMPLEX);
+            cv::rectangle(rectifiedRight, cv::Rect(cv::Point(x1, y1), cv::Point(x2, y2)), color, cv::FONT_HERSHEY_SIMPLEX);
         }
 
         std::stringstream fpsStr;
         fpsStr << std::fixed << std::setprecision(2) << fps;
-        cv::putText(frame, fpsStr.str(), cv::Point(2, imgFrame->getHeight() - 4), cv::FONT_HERSHEY_TRIPLEX, 0.4, color);
+        cv::putText(rectifiedRight, fpsStr.str(), cv::Point(2, rectifiedRight.rows - 4), cv::FONT_HERSHEY_TRIPLEX, 0.4, color);
 
         cv::imshow("depth", depthFrameColor);
-        cv::imshow("preview", frame);
+        cv::imshow("rectified right", rectifiedRight);
         int key = cv::waitKey(1);
         if(key == 'q') {
             return 0;
