+++ examples/src/27_spatial_location_calculator.cpp
