+++ examples/src/28_2_mono_camera_video.cpp
