+++ examples/src/30_2_stereo_depth_video.cpp
